backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai>=1.12.0
yt-dlp>=2024.1.0
ffmpeg-python==0.2.0
aiofiles==23.2.1
orjson==3.9.10
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx>=0.25.0

